break time working time of the day ui done
