+++ app/src/main/java/com/example/myapplication/TimeOfTheDay.java
@@ -1,0 +1,12 @@
+package com.example.myapplication;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class TimeOfTheDay extends AppCompatActivity {
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_time_of_the_day);
+    }
+}
