+++ app/src/main/res/drawable/btn_gradient.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+
+
+    <gradient
+
+        android:startColor="#00cc96"
+        android:endColor="#00cc96"
+
+
+        />
+
+
+    <corners
+
+        android:radius="20dp"
+
+        />
+
+
+
+
+
+</shape>
