+++ app/src/main/res/layout/activity_time_of_the_day.xml
@@ -163,9 +163,7 @@
             android:layout_marginEnd="20dp"
             android:layout_marginRight="20dp"
             android:layout_weight="1"
-            android:background="@drawable/btn_gradient"
             android:text="set"
-            app:backgroundTint="?android:attr/textColorHint"
             app:backgroundTintMode="add" />
 
         <Button
@@ -178,9 +176,7 @@
             android:layout_marginEnd="70dp"
             android:layout_marginRight="70dp"
             android:layout_weight="1"
-            android:background="@drawable/btn_gradient"
             android:text="new"
-            app:backgroundTint="?android:attr/textColorHint"
             app:backgroundTintMode="add" />
     </LinearLayout>
