+++ app/src/main/res/layout/home.xml
@@ -1,0 +1,84 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    tools:context=".BreakTime" >
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal">
+
+        <TextView
+            android:id="@+id/logo_get_off"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="30dp"
+            android:layout_marginLeft="30dp"
+            android:layout_marginTop="20dp"
+            android:layout_weight="1"
+            android:text="Get-Off"
+            android:textColor="#00CC96"
+            android:textSize="30sp"
+            android:textStyle="bold" />
+
+        <ImageView
+            android:id="@+id/imageView6"
+            android:layout_width="5dp"
+            android:layout_height="30dp"
+            android:layout_marginTop="23dp"
+            android:layout_weight="1"
+            app:srcCompat="@drawable/verticaldots" />
+
+    </LinearLayout>
+
+    <Button
+        android:id="@+id/break_time_button"
+        android:layout_width="match_parent"
+        android:layout_height="150dp"
+        android:layout_marginStart="35dp"
+        android:layout_marginLeft="35dp"
+        android:layout_marginTop="100dp"
+        android:layout_marginEnd="35dp"
+        android:layout_marginRight="35dp"
+        android:layout_marginBottom="20dp"
+        android:text="break time"
+        android:textSize="25sp"
+        android:textStyle="bold"
+        app:backgroundTintMode="add"
+        app:cornerRadius="50dp" />
+
+    <Button
+        android:id="@+id/time_of_the_day_button"
+        android:layout_width="match_parent"
+        android:layout_height="150dp"
+        android:layout_marginStart="35dp"
+        android:layout_marginLeft="35dp"
+        android:layout_marginEnd="35dp"
+        android:layout_marginRight="35dp"
+        android:layout_marginBottom="20dp"
+        android:text="time of the day"
+        android:textSize="25sp"
+        android:textStyle="bold"
+        app:backgroundTintMode="add"
+        app:cornerRadius="50dp" />
+
+    <Button
+        android:id="@+id/app_specific_button"
+        android:layout_width="match_parent"
+        android:layout_height="150dp"
+        android:layout_marginStart="35dp"
+        android:layout_marginLeft="35dp"
+        android:layout_marginEnd="35dp"
+        android:layout_marginRight="35dp"
+        android:layout_marginBottom="20dp"
+        android:text="app specific"
+        android:textSize="25sp"
+        android:textStyle="bold"
+        app:backgroundTintMode="add"
+        app:cornerRadius="50dp" />
+
+</LinearLayout>
