+++ app/src/main/res/values/colors.xml
@@ -7,7 +7,7 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
-    <color name="colorPrimary">#67C294</color>
+    <color name="colorPrimary">#00cc96</color>
     <color name="colorPrimaryVariant">#009F88</color>
     <color name="colorOnPrimary">#E2EFF1</color>
 </resources>
